plop gh completions in place; need to run compinit for fnm in fpath :(

+++ zsh/.zshrc
@@ -20,31 +20,53 @@ alias rm="nocorrect rm"
 alias mkdir="nocorrect mkdir"
 
 # ============================================================================
+# Modules: compinit and more completion
+# ============================================================================
+
+compdir="${LDOTDIR}/completions"
+mkdir -pv "$compdir"
+
+__dko_has 'fnm' &&
+  [ ! -f "${compdir}/_fnm" ] &&
+  fnm completions --shell zsh > "${compdir}/_fnm"
+
+# ============================================================================
 # zinit
 # ============================================================================
 
-__dko_has 'git' && {
+# file is required and missing on busybox
+__dko_has 'file' && __dko_has 'git' && {
   declare -A ZINIT
   ZINIT[HOME_DIR]="${XDG_DATA_HOME}/zinit"
 
   # part of zinit's install, found by compaudit
-  mkdir -p "${ZINIT[HOME_DIR]}" && chmod g-rwX "${ZINIT[HOME_DIR]}"
+  mkdir -pv "${ZINIT[HOME_DIR]}" && chmod g-rwX "${ZINIT[HOME_DIR]}"
 
-  dko_zinit_dest="${ZINIT[HOME_DIR]}/bin"
-  dko_zinit_script="${dko_zinit_dest}/zinit.zsh"
-  . "$dko_zinit_script" 2>/dev/null || {
-    # install if needed
-    command git clone https://github.com/zdharma/zinit "${dko_zinit_dest}" &&
-      . "$dko_zinit_script"
+  function {
+    local zinit_dest="${ZINIT[HOME_DIR]}/bin"
+    local zinit_script="${zinit_dest}/zinit.zsh"
+    . "$zinit_script" 2>/dev/null || {
+      # install if needed
+      command git clone https://github.com/zdharma/zinit "${zinit_dest}" &&
+        . "$zinit_script"
+    }
   }
-  unset dko_zinit_dest
-  unset dko_zinit_script
+}
 
+if __dko_has 'zinit'; then
+  zinit add-fpath "$compdir"
   . "${ZDOTDIR}/zinit.zsh" 2>/dev/null && {
     autoload -Uz _zinit && (( ${+_comps} )) && _comps[zinit]=_zinit
     alias unzinit='rm -rf "${ZINIT[HOME_DIR]}"'
   }
-}
+else
+  fpath+=($compdir)
+fi
+
+unset compdir
+
+autoload -Uz compinit && compinit
+compdef g=git
 
 # ============================================================================
 # Options
@@ -104,9 +126,6 @@ setopt VI
 # ============================================================================
 # Modules
 # ============================================================================
-
-! __dko_has 'zinit' && autoload -Uz compinit && compinit
-compdef g=git
 
 # color complist
 zmodload -i zsh/complist
@@ -219,7 +238,7 @@ bindkey '^w' vi-forward-word
 # FZF keybindings
 # ============================================================================
 
-if __dko_has "fzf"; then
+if __dko_has 'fzf'; then
   if . "${XDG_CONFIG_HOME}/fzf/fzf.zsh" 2>/dev/null || {
     # linux package managers throw it here
     . "/usr/share/fzf/completion.zsh" 2>/dev/null
